Completed getUserList() for ClientNode

+++ etapa3PI/Intermediary.hpp
@@ -7,7 +7,7 @@ private:
   bool handleDatagram(int client_socket, char* datagram
     , size_t datagram_size) override;
   // Método que se encarga de reenviar un mensaje largo
-  bool resendLongString(int source_socket, int dest_socket, size_t total_size, int timeout_sec);
+  bool resendLongString(int source_socket, int dest_socket, int total_size, int timeout_sec);
 
 public:
   Intermediary(std::string logFilename, int processId, int port);
